fix: move Drug to pharmacy.js and use updateBenefitValue() instead of updateDrugs() to not break public API of classes
